seperated out TeamDisplay styles and functions
